+++ packages/pglite/scripts/bundle-wasm.ts
@@ -83,6 +83,7 @@ async function main() {
     `require("./postgres.cjs").default`,
     ['.cjs'],
   )
+  await findAndReplaceInDir('./dist/pg_hashids', /\.\.\/release\//g, '', ['.js', '.cjs'])
 }
 
 await main()