+++ packages/pglite/tsup.config.ts
@@ -29,6 +29,7 @@ const entryPoints = [
   'src/pg_uuidv7/index.ts',
   'src/age/index.ts',
   'src/worker/index.ts',
+  'src/pg_hashids/index.ts',
 ]
 
 const contribDir = path.join(root, 'src', 'contrib')